Fixed read metrics: incrementally update read metrics.
This fixes a problem where the read metrics are recorded as the
full input split size, even when the task doesn't read the entire
input split.

+++ core/src/main/scala/org/apache/spark/executor/TaskMetrics.scala
@@ -154,7 +154,7 @@ case class InputMetrics(val readMethod: DataReadMethod.Value) {
   /**
    * Packets (just for debugging).
    */
-  var numPackets: Int = 0
+  var numPackets: Long = 0L
 
   /**
    * Time Hadoop spent opening the file for reading.
